exercio 36 renomada classe reescrito codigo - treino
